+++ utils.py
@@ -11,7 +11,7 @@ Auteur        : Pierre Théberge
 Compagnie     : Innovations, Performances, Technologies inc.
 Créé le       : 2025-08-05
 Modifié le    : 2026-09-24
-Version       : 0.6.0
+Version       : 0.6.1
 Copyright     : Pierre Théberge
 
 Description
@@ -114,6 +114,7 @@ Modifications
 0.5.25 - 2026-09-23   CR      : Synchronisation de version (aucun changement fonctionnel).
 0.6.0  - 2026-09-24   ES-28   : get_last_downloaded_report_file : paramètre depuis ; pause_on_error
                                 n'attend que si stdin est un terminal et la session interactive.
+0.6.1  - 2026-09-24   ES-28   : Synchronisation de version (aucun changement fonctionnel).
 
 Paramètres
 ----------